refactor(file-structure): standardize controller file locations to module root
Move controllers from subdirectories to module root per NestJS CLI conventions.
This aligns the codebase with standard NestJS file structure where controllers
are located at module root unless there are compelling organizational reasons.

Changes:
- Move trackers controllers to module root
- Move mmr-calculation controllers to module root
- Move internal controller to module root
- Remove dead code (unused CalculatorController)
- Update all import paths
- Create FILE_STRUCTURE.md documentation
- Clean up empty controller subdirectories

Refs: #145
