+++ src/internal/internal.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { InternalController } from './internal.controller';
 import { InternalTrackerController } from './internal-tracker.controller';
-import { InternalScheduledProcessingController } from './controllers/internal-scheduled-processing.controller';
+import { InternalScheduledProcessingController } from './internal-scheduled-processing.controller';
 import { TrackersModule } from '../trackers/trackers.module';
 
 @Module({
