+++ src/trackers/tracker-admin.controller.ts
@@ -16,14 +16,14 @@ import {
   ApiParam,
   ApiQuery,
 } from '@nestjs/swagger';
-import { JwtAuthGuard } from '../../auth/guards/jwt-auth.guard';
-import { SystemAdminGuard } from '../../common/authorization/guards/system-admin/system-admin.guard';
-import { TrackerService } from '../tracker.service';
-import { TrackerProcessingService } from '../services/tracker-processing.service';
-import { TrackerRefreshSchedulerService } from '../services/tracker-refresh-scheduler.service';
-import { BatchRefreshDto } from '../dto/batch-refresh.dto';
+import { JwtAuthGuard } from '../auth/guards/jwt-auth.guard';
+import { SystemAdminGuard } from '../common/authorization/guards/system-admin/system-admin.guard';
+import { TrackerService } from './tracker.service';
+import { TrackerProcessingService } from './services/tracker-processing.service';
+import { TrackerRefreshSchedulerService } from './services/tracker-refresh-scheduler.service';
+import { BatchRefreshDto } from './dto/batch-refresh.dto';
 import { TrackerScrapingStatus, GamePlatform } from '@prisma/client';
-import { ParseCUIDPipe, ParseEnumPipe } from '../../common/pipes';
+import { ParseCUIDPipe, ParseEnumPipe } from '../common/pipes';
 
 @ApiTags('Admin - Trackers')
 @Controller('api/admin/trackers')
